Bug 1172264 - Switch MediaDecoder's mDuration represenation to a double. r=jww

+++ dom/media/MediaDecoder.h
@@ -943,10 +943,8 @@ public:
   AbstractCanonical<bool>* CanonicalPreservesPitch() { return &mPreservesPitch; }
 protected:
 
-  // Duration of the media resource. Set to -1 if unknown.
-  // Set when the metadata is loaded. Accessed on the main thread
-  // only.
-  int64_t mDuration;
+  // Official duration of the media resource as observed by script.
+  double mDuration;
 
   // True if the media is seekable (i.e. supports random access).
   bool mMediaSeekable;
